Create Dynamic World_Land cover_Benchmark
📡 GEE Dynamic World 10m vs. ESA WorldCover 10m.
- Exercise over Big Island of Chiloe, Southern Chile.-

When conducting a Land Use Assessment (LLU), the free and readily available dataset until a few years ago was ESA WorldCover 10m. However, the introduction of Dynamic World changed the game, not only because of its ability to reflect information with a 10-meter resolution (the same as ESA WorldCover) but also, and this is my opinion, because of its high update frequency of 6 days. This capacity is particularly evident in urban areas, which tend to rapidly impact their surroundings, making constant updates essential. In this respect, Dynamic World is comparatively the best free option available to everyone.